fixing environment for swirl

+++ server_apps/DB_maintenance/postgres/runPostgresBuild.sh
@@ -9,7 +9,10 @@ $INFORMIXDIR/bin/dbaccess $DBNAME $SOURCEROOT/server_apps/DB_maintenance/postgre
 
 echo "drop and recreate database in $DBNAME value"
 
-$PGBINDIR/dropdb $DBNAME
+if $PGBINDIR/psql -lqt | cut -d \| -f 1 | grep -qw $DBNAME; then
+    $PGBINDIR/dropdb $DBNAME
+fi
+
 $PGBINDIR/createdb $DBNAME
 
 echo "****Build liquibase changelog"
@@ -22,8 +25,9 @@ rm -rf /tmp/changelogMigrationFile.xml
 rm -rf $SOURCEROOT/source/org/zfin/db/postgres/changelogMigrationFile.xml
 rm -rf $SOURCEROOT/source/org/zfin/db/postgres/constraints/changelogConstraintFile.xml
 rm -rf $SOURCEROOT/source/org/zfin/db/postgres/tableMigration.xml
+rm -rf /tmp/liquibaseOutput.txt
 
-./liquibase --driver=com.informix.jdbc.IfxDriver --url="jdbc:informix-sqli://${HOSTNAME}:${INFORMIX_PORT}/${DBNAME}:INFORMIXSERVER=${INFORMIXSERVER};DB_LOCALE=en_US.utf8" --defaultSchemaName="informix" --classpath="$SOURCEROOT/lib/Java/informix/ifxjdbc-3.70.JC1.jar" --changeLogFile="/tmp/changelogMigrationFile.xml" generateChangeLog
+./liquibase --driver=com.informix.jdbc.IfxDriver --url="jdbc:informix-sqli://${HOSTNAME}:${INFORMIX_PORT}/${DBNAME}:INFORMIXSERVER=${INFORMIXSERVER};DB_LOCALE=en_US.utf8" --defaultSchemaName="informix" --classpath="$SOURCEROOT/lib/Java/informix/ifxjdbc-3.70.JC1.jar" --changeLogFile="/tmp/changelogMigrationFile.xml" generateChangeLog > /tmp/liquibaseOutput.txt
 
 cp /tmp/changelogMigrationFile.xml $SOURCEROOT/source/org/zfin/db/postgres/changelogMigrationFile.xml
